refactored code into submitWork.js
