+++ src/api.v2/helpers/worker/workSubmit/submitMarkerHeatmapWork.js
@@ -1,15 +1,10 @@
-const createObjectHash = require('../createObjectHash');
-const config = require('../../../../config');
-const validateAndSubmitWork = require('../../../events/validateAndSubmitWork');
-const getExperimentBackendStatus = require('../../backendStatus/getExperimentBackendStatus');
 const getExtraDependencies = require('./getExtraDependencies');
+const submitWork = require('./submitWork');
 
 
 const submitMarkerHeatmapWork = async (message) => {
   const { experimentId, input: { authJWT } } = message;
 
-  const backendStatus = await getExperimentBackendStatus(experimentId);
-  const { pipeline: { startDate: qcPipelineStartDate } } = backendStatus;
   const numGenes = 5;
   const selectedCellSet = 'louvain';
 
@@ -19,39 +14,10 @@ const submitMarkerHeatmapWork = async (message) => {
     cellSetKey: selectedCellSet,
   };
 
-  const cacheUniquenessKey = null;
 
-  const extras = undefined;
   const extraDependencies = await getExtraDependencies(body.name, message);
-  const { workerVersion } = config;
-  const ETagBody = {
-    experimentId,
-    body,
-    qcPipelineStartDate: qcPipelineStartDate.toISOString(),
-    extras,
-    cacheUniquenessKey,
-    workerVersion,
-    extraDependencies,
-  };
+  const ETag = await submitWork(experimentId, authJWT, body, extraDependencies);
 
-
-  const ETag = createObjectHash(ETagBody);
-  const now = new Date();
-  const timeout = 15 * 60 * 1000; // 15min in ms
-  const timeoutDate = new Date(now.getTime() + timeout);
-  const request = {
-    ETag,
-    socketId: 'randomID',
-    experimentId,
-    authJWT,
-    timeout: timeoutDate.toISOString(),
-    body,
-  };
-
-  await validateAndSubmitWork(request);
-
-  console.log('markerHeatmap - body: ', ETagBody);
-  console.log('markerHeatmap - etag: ', ETag);
   // explicitly return ETag to make it stand out more in tests and so harder to break
   return ETag;
 };
